+++ src/main/java/blackjack/domain/card/Deck.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import java.util.*;
 import java.util.stream.Collectors;
