+++ src/test/java/blackjack/domain/user/NameTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.user;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
